Change buttons in bill popUp for better UI
Fix notifications Issue
