+++ lib/src/ui/screens/home/HomePage.dart
@@ -72,11 +72,6 @@ class HomePage extends StatefulWidget {
   var destination;
 
   CustomerModel? customer;
-
-  var samePositionCount = 0;
-
-  bool? hasGps = false;
-
   HomePage(
       {Key? key,
       this.destination,
@@ -107,8 +102,6 @@ class _HomePageState extends State<HomePage> {
 
   var subscription;
 
-  StreamSubscription<Position>? positionStream;
-  Position? tmpLocation;
 
   Future<int> checkLogin() async {
     StatefulWidget launchPage =
@@ -204,10 +197,6 @@ class _HomePageState extends State<HomePage> {
   void initState() {
    // get_token();
 
-
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      _getLastKnowLocation(jumpToBuyPageDetails: false);
-    });
     homeWelcomePage = HomeWelcomeNewPage(
         key: homeKey,
         presenter: HomeWelcomePresenter(HomeWelcomeView()),
@@ -764,7 +753,6 @@ class _HomePageState extends State<HomePage> {
         }
       }
     } else if (value == 1) {
-      _getLastKnowLocation(jumpToBuyPageDetails: true);
     } else {
       // 0
       setState(() {
@@ -1137,282 +1125,6 @@ class _HomePageState extends State<HomePage> {
     }
   }
 
-  Future _getLastKnowLocation({bool jumpToBuyPageDetails = false}) async {
-    SharedPreferences.getInstance().then((value) async {
-      prefs = value;
-
-      String? _has_accepted_gps = await prefs.getString("_has_accepted_gps");
-      /* no need to commit */
-      /* expiration date in 3months */
-      if (_has_accepted_gps != "ok") {
-        return showDialog<void>(
-          context: context,
-          barrierDismissible: false, // user must tap button!
-          builder: (BuildContext context) {
-            return AlertDialog(
-              title: Text(
-                  "${AppLocalizations.of(context)!.translate('request')}"
-                      .toUpperCase(),
-                  style: TextStyle(color: KColors.primaryColor)),
-              content: SingleChildScrollView(
-                child: ListBody(
-                  children: <Widget>[
-                    // location_permission
-                    Container(
-                        height: 100,
-                        width: 100,
-                        decoration: BoxDecoration(
-                            image: new DecorationImage(
-                          image: new AssetImage(ImageAssets.address),
-                        ))),
-                    SizedBox(height: 10),
-                    Text(
-                        "${AppLocalizations.of(context)!.translate('location_explanation_pricing')}",
-                        textAlign: TextAlign.center,
-                        style: TextStyle(fontSize: 14))
-                  ],
-                ),
-              ),
-              actions: <Widget>[
-                TextButton(
-                  child: Text(
-                      "${AppLocalizations.of(context)!.translate('refuse')}"),
-                  onPressed: () {
-                    Navigator.of(context).pop();
-                  },
-                ),
-                TextButton(
-                  child: Text(
-                      "${AppLocalizations.of(context)!.translate('accept')}"),
-                  onPressed: () {
-                    prefs!.setString("_has_accepted_gps", "ok");
-                    // call get location again...
-                    Future.delayed(Duration(milliseconds: 1000), () {
-                      _getLastKnowLocation(
-                          jumpToBuyPageDetails: jumpToBuyPageDetails);
-                    });
-                    Navigator.of(context).pop();
-                  },
-                )
-              ],
-            );
-          },
-        );
-      } else {
-        // permission has been accepted
-        LocationPermission permission = await Geolocator.checkPermission();
-        if (permission == LocationPermission.deniedForever) {
-          /*  ---- */
-          // await Geolocator.openAppSettings();
-          /* ---- */
-          return showDialog<void>(
-            context: context,
-            barrierDismissible: false, // user must tap button!
-            builder: (BuildContext context) {
-              return AlertDialog(
-                title: Text(
-                    "${AppLocalizations.of(context)!.translate('permission_')}"
-                        .toUpperCase(),
-                    style: TextStyle(color: KColors.primaryColor)),
-                content: SingleChildScrollView(
-                  child: ListBody(
-                    children: <Widget>[
-                      /* add an image*/
-                      // location_permission
-                      Container(
-                          height: 100,
-                          width: 100,
-                          decoration: BoxDecoration(
-                              image: new DecorationImage(
-                            fit: BoxFit.fitHeight,
-                            image: new AssetImage(ImageAssets.address),
-                          ))),
-                      SizedBox(height: 10),
-                      Text(
-                          "${AppLocalizations.of(context)!.translate('request_location_permission')}",
-                          textAlign: TextAlign.center,
-                          style: TextStyle(fontSize: 14))
-                    ],
-                  ),
-                ),
-                actions: <Widget>[
-                  TextButton(
-                    child: Text(
-                        "${AppLocalizations.of(context)!.translate('refuse')}"),
-                    onPressed: () {
-                      Navigator.of(context).pop();
-                    },
-                  ),
-                  TextButton(
-                    child: Text(
-                        "${AppLocalizations.of(context)!.translate('accept')}"),
-                    onPressed: () async {
-                      /* */
-                      await Geolocator.openAppSettings();
-                      Navigator.of(context).pop();
-                    },
-                  )
-                ],
-              );
-            },
-          );
-          /* ---- */
-        } else if (permission == LocationPermission.denied) {
-          /* ---- */
-          // Geolocator.requestPermission();
-          /* ---- */
-          return showDialog<void>(
-            context: context,
-            barrierDismissible: false,
-            builder: (BuildContext context) {
-              return AlertDialog(
-                title: Text(
-                    "${AppLocalizations.of(context)!.translate('permission_')}"
-                        .toUpperCase(),
-                    style: TextStyle(color: KColors.primaryColor)),
-                content: SingleChildScrollView(
-                  child: ListBody(
-                    children: <Widget>[
-                      /* add an image*/
-                      // location_permission
-                      Container(
-                          height: 100,
-                          width: 100,
-                          decoration: BoxDecoration(
-                              image: new DecorationImage(
-                            fit: BoxFit.fitHeight,
-                            image: new AssetImage(ImageAssets.address),
-                          ))),
-                      SizedBox(height: 10),
-                      Text(
-                          "${AppLocalizations.of(context)!.translate('request_location_permission')}",
-                          textAlign: TextAlign.center,
-                          style: TextStyle(fontSize: 14))
-                    ],
-                  ),
-                ),
-                actions: <Widget>[
-                  TextButton(
-                    child: Text(
-                        "${AppLocalizations.of(context)!.translate('refuse')}"),
-                    onPressed: () {
-                      Navigator.of(context).pop();
-                    },
-                  ),
-                  TextButton(
-                    child: Text(
-                        "${AppLocalizations.of(context)!.translate('accept')}"),
-                    onPressed: () async {
-                      /* */
-                      await Geolocator.requestPermission();
-                      LocationPermission permission2 =
-                          await Geolocator.checkPermission();
-                      if (permission2 == LocationPermission.always ||
-                          permission2 == LocationPermission.whileInUse) {
-                        _getLastKnowLocation(
-                            jumpToBuyPageDetails: jumpToBuyPageDetails);
-                      }
-                      Navigator.of(context).pop();
-                    },
-                  )
-                ],
-              );
-            },
-          );
-        } else {
-          bool isLocationServiceEnabled =
-              await Geolocator.isLocationServiceEnabled();
-          if (!isLocationServiceEnabled) {
-            return showDialog<void>(
-              context: context,
-              barrierDismissible: false, // user must tap button!
-              builder: (BuildContext context) {
-                return AlertDialog(
-                  title: Text(
-                      "${AppLocalizations.of(context)!.translate('permission_')}"
-                          .toUpperCase(),
-                      style: TextStyle(color: KColors.primaryColor)),
-                  content: SingleChildScrollView(
-                    child: ListBody(
-                      children: <Widget>[
-                        Container(
-                            height: 100,
-                            width: 100,
-                            decoration: BoxDecoration(
-                                image: new DecorationImage(
-                              fit: BoxFit.fitHeight,
-                              image: new AssetImage(
-                                  ImageAssets.location_permission),
-                            ))),
-                        SizedBox(height: 10),
-                        Text(
-                            "${AppLocalizations.of(context)!.translate('request_location_activation_permission')}",
-                            textAlign: TextAlign.center,
-                            style: TextStyle(fontSize: 14))
-                      ],
-                    ),
-                  ),
-                  actions: <Widget>[
-                    TextButton(
-                      child: Text(
-                          "${AppLocalizations.of(context)!.translate('refuse')}"),
-                      onPressed: () {
-                        Navigator.of(context).pop();
-                      },
-                    ),
-                    TextButton(
-                      child: Text(
-                          "${AppLocalizations.of(context)!.translate('accept')}"),
-                      onPressed: () async {
-                        /* */
-                        Navigator.of(context).pop();
-                        await Geolocator.openLocationSettings();
-                      },
-                    )
-                  ],
-                );
-              },
-            );
-            /* ---- */
-          } else {
-            /* show loading dialog until this finishes then close */
-
-            // switch to page two
-            if (jumpToBuyPageDetails) {
-              setState(() {
-                StateContainer.of(context).updateTabPosition(tabPosition: 1);
-              });
-            }
-
-            positionStream =
-                Geolocator.getPositionStream().listen((Position position) {
-              /* compare current and old position */
-              if (position?.latitude != null &&
-                  tmpLocation?.latitude != null &&
-                  (position.latitude * 100).round() ==
-                      (tmpLocation!.latitude! * 100).round() &&
-                  (position.longitude * 100).round() ==
-                      (tmpLocation!.longitude * 100).round()) {
-                widget.samePositionCount++;
-              } else {
-                widget.samePositionCount = 0;
-                tmpLocation = StateContainer.of(context).location;
-                if (position != null && mounted) {
-                  widget.hasGps = true;
-                  setState(() {
-                    StateContainer.of(context)
-                        .updateLocation(location: position);
-                  });
-                }
-              }
-              if (widget.samePositionCount >= 3 || widget.hasGps!)
-                positionStream?.cancel();
-            });
-          }
-        }
-      }
-    });
-  }
 
   void _requestGpsPermissionAndLocation() {
     /* has been requested already, we shouldnt request a second time during this time */
@@ -1421,13 +1133,7 @@ class _HomePageState extends State<HomePage> {
       StateContainer.of(context).location_asked = true;
     else
       return;
-    if (mounted) {
-      _getLastKnowLocation();
-      if (widget.hasGps == false &&
-          StateContainer?.of(context)?.location != null) {
-        xrint("init -- 1");
-      } else {}
-    }
+
   }
 }
 
@@ -1457,6 +1163,7 @@ NotificationItem? _notificationFromMessage(Map<String, dynamic> messageEntry) {
 Future<void> iLaunchNotifications(NotificationItem notificationItem) async {
   String groupKey = "tg.tmye.kaba.brave.one";
   final String? bigPictureUrl = notificationItem.image_link?.toString();
+
   String? filePath;
   if (bigPictureUrl != null && bigPictureUrl.isNotEmpty) {
     try {
