+++ lib/src/ui/screens/home/buy/main/ServiceMainPage.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:io';
 
 import 'package:KABA/src/StateContainer.dart';
@@ -65,7 +66,8 @@ class ServiceMainPage extends StatefulWidget {
   List<ServiceMainEntity>? coming_soon_services = [];
 
   Position? initialLocation;
-
+  var samePositionCount = 0;
+  bool? hasGps = false;
   ServiceMainPage({Key? key, this.presenter}) : super(key: key);
 
   @override
@@ -87,11 +89,15 @@ class ServiceMainPageState extends State<ServiceMainPage>
   bool isPickLocation = false;
 
   CurrentLocationTile? _myCurrentTile;
+  StreamSubscription<Position>? positionStream;
+  Position? tmpLocation;
 
   @override
   void initState() {
     super.initState();
     this.widget.presenter!.checkVersion();
+
+
     widget.presenter!.serviceMainView = this;
 
     if (widget.available_services == null) widget.available_services = [];
@@ -141,7 +147,13 @@ class ServiceMainPageState extends State<ServiceMainPage>
         CustomerUtils utils = CustomerUtils();
         bool isUpdateSeen = await utils.getViewUpdate();
         xrint("isUpdateSeen $isUpdateSeen");
-        if(!isUpdateSeen){showNewFeature(context, code);}
+        if(!isUpdateSeen){
+          showNewFeature(context, code);
+        }else{
+          WidgetsBinding.instance.addPostFrameCallback((_) {
+            _getLastKnowLocation(jumpToBuyPageDetails: false);
+          });
+        }
       }
     });
   }
@@ -188,6 +200,9 @@ class ServiceMainPageState extends State<ServiceMainPage>
                     CustomerUtils utils = CustomerUtils();
                     await utils.setViewUpdate(enable: true);
                     overlayEntry!.remove();
+                    WidgetsBinding.instance.addPostFrameCallback((_) {
+                      _getLastKnowLocation(jumpToBuyPageDetails: false);
+                    });
                   },
                 )
               ],
@@ -1184,6 +1199,282 @@ class ServiceMainPageState extends State<ServiceMainPage>
 
     setState(() {
       isPickLocation = false;
+    });
+  }
+  Future _getLastKnowLocation({bool jumpToBuyPageDetails = false}) async {
+    SharedPreferences.getInstance().then((value) async {
+      prefs = value;
+
+      String? _has_accepted_gps = await prefs.getString("_has_accepted_gps");
+      /* no need to commit */
+      /* expiration date in 3months */
+      if (_has_accepted_gps != "ok") {
+        return showDialog<void>(
+          context: context,
+          barrierDismissible: false, // user must tap button!
+          builder: (BuildContext context) {
+            return AlertDialog(
+              title: Text(
+                  "${AppLocalizations.of(context)!.translate('request')}"
+                      .toUpperCase(),
+                  style: TextStyle(color: KColors.primaryColor)),
+              content: SingleChildScrollView(
+                child: ListBody(
+                  children: <Widget>[
+                    // location_permission
+                    Container(
+                        height: 100,
+                        width: 100,
+                        decoration: BoxDecoration(
+                            image: new DecorationImage(
+                              image: new AssetImage(ImageAssets.address),
+                            ))),
+                    SizedBox(height: 10),
+                    Text(
+                        "${AppLocalizations.of(context)!.translate('location_explanation_pricing')}",
+                        textAlign: TextAlign.center,
+                        style: TextStyle(fontSize: 14))
+                  ],
+                ),
+              ),
+              actions: <Widget>[
+                TextButton(
+                  child: Text(
+                      "${AppLocalizations.of(context)!.translate('refuse')}"),
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                  },
+                ),
+                TextButton(
+                  child: Text(
+                      "${AppLocalizations.of(context)!.translate('accept')}"),
+                  onPressed: () {
+                    prefs!.setString("_has_accepted_gps", "ok");
+                    // call get location again...
+                    Future.delayed(Duration(milliseconds: 1000), () {
+                      _getLastKnowLocation(
+                          jumpToBuyPageDetails: jumpToBuyPageDetails);
+                    });
+                    Navigator.of(context).pop();
+                  },
+                )
+              ],
+            );
+          },
+        );
+      } else {
+        // permission has been accepted
+        LocationPermission permission = await Geolocator.checkPermission();
+        if (permission == LocationPermission.deniedForever) {
+          /*  ---- */
+          // await Geolocator.openAppSettings();
+          /* ---- */
+          return showDialog<void>(
+            context: context,
+            barrierDismissible: false, // user must tap button!
+            builder: (BuildContext context) {
+              return AlertDialog(
+                title: Text(
+                    "${AppLocalizations.of(context)!.translate('permission_')}"
+                        .toUpperCase(),
+                    style: TextStyle(color: KColors.primaryColor)),
+                content: SingleChildScrollView(
+                  child: ListBody(
+                    children: <Widget>[
+                      /* add an image*/
+                      // location_permission
+                      Container(
+                          height: 100,
+                          width: 100,
+                          decoration: BoxDecoration(
+                              image: new DecorationImage(
+                                fit: BoxFit.fitHeight,
+                                image: new AssetImage(ImageAssets.address),
+                              ))),
+                      SizedBox(height: 10),
+                      Text(
+                          "${AppLocalizations.of(context)!.translate('request_location_permission')}",
+                          textAlign: TextAlign.center,
+                          style: TextStyle(fontSize: 14))
+                    ],
+                  ),
+                ),
+                actions: <Widget>[
+                  TextButton(
+                    child: Text(
+                        "${AppLocalizations.of(context)!.translate('refuse')}"),
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                    },
+                  ),
+                  TextButton(
+                    child: Text(
+                        "${AppLocalizations.of(context)!.translate('accept')}"),
+                    onPressed: () async {
+                      /* */
+                      await Geolocator.openAppSettings();
+                      Navigator.of(context).pop();
+                    },
+                  )
+                ],
+              );
+            },
+          );
+          /* ---- */
+        } else if (permission == LocationPermission.denied) {
+          /* ---- */
+          // Geolocator.requestPermission();
+          /* ---- */
+          return showDialog<void>(
+            context: context,
+            barrierDismissible: false,
+            builder: (BuildContext context) {
+              return AlertDialog(
+                title: Text(
+                    "${AppLocalizations.of(context)!.translate('permission_')}"
+                        .toUpperCase(),
+                    style: TextStyle(color: KColors.primaryColor)),
+                content: SingleChildScrollView(
+                  child: ListBody(
+                    children: <Widget>[
+                      /* add an image*/
+                      // location_permission
+                      Container(
+                          height: 100,
+                          width: 100,
+                          decoration: BoxDecoration(
+                              image: new DecorationImage(
+                                fit: BoxFit.fitHeight,
+                                image: new AssetImage(ImageAssets.address),
+                              ))),
+                      SizedBox(height: 10),
+                      Text(
+                          "${AppLocalizations.of(context)!.translate('request_location_permission')}",
+                          textAlign: TextAlign.center,
+                          style: TextStyle(fontSize: 14))
+                    ],
+                  ),
+                ),
+                actions: <Widget>[
+                  TextButton(
+                    child: Text(
+                        "${AppLocalizations.of(context)!.translate('refuse')}"),
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                    },
+                  ),
+                  TextButton(
+                    child: Text(
+                        "${AppLocalizations.of(context)!.translate('accept')}"),
+                    onPressed: () async {
+                      /* */
+                      await Geolocator.requestPermission();
+                      LocationPermission permission2 =
+                      await Geolocator.checkPermission();
+                      if (permission2 == LocationPermission.always ||
+                          permission2 == LocationPermission.whileInUse) {
+                        _getLastKnowLocation(
+                            jumpToBuyPageDetails: jumpToBuyPageDetails);
+                      }
+                      Navigator.of(context).pop();
+                    },
+                  )
+                ],
+              );
+            },
+          );
+        } else {
+          bool isLocationServiceEnabled =
+          await Geolocator.isLocationServiceEnabled();
+          if (!isLocationServiceEnabled) {
+            return showDialog<void>(
+              context: context,
+              barrierDismissible: false, // user must tap button!
+              builder: (BuildContext context) {
+                return AlertDialog(
+                  title: Text(
+                      "${AppLocalizations.of(context)!.translate('permission_')}"
+                          .toUpperCase(),
+                      style: TextStyle(color: KColors.primaryColor)),
+                  content: SingleChildScrollView(
+                    child: ListBody(
+                      children: <Widget>[
+                        Container(
+                            height: 100,
+                            width: 100,
+                            decoration: BoxDecoration(
+                                image: new DecorationImage(
+                                  fit: BoxFit.fitHeight,
+                                  image: new AssetImage(
+                                      ImageAssets.location_permission),
+                                ))),
+                        SizedBox(height: 10),
+                        Text(
+                            "${AppLocalizations.of(context)!.translate('request_location_activation_permission')}",
+                            textAlign: TextAlign.center,
+                            style: TextStyle(fontSize: 14))
+                      ],
+                    ),
+                  ),
+                  actions: <Widget>[
+                    TextButton(
+                      child: Text(
+                          "${AppLocalizations.of(context)!.translate('refuse')}"),
+                      onPressed: () {
+                        Navigator.of(context).pop();
+                      },
+                    ),
+                    TextButton(
+                      child: Text(
+                          "${AppLocalizations.of(context)!.translate('accept')}"),
+                      onPressed: () async {
+                        /* */
+                        Navigator.of(context).pop();
+                        await Geolocator.openLocationSettings();
+                      },
+                    )
+                  ],
+                );
+              },
+            );
+            /* ---- */
+          } else {
+            /* show loading dialog until this finishes then close */
+
+            // switch to page two
+            if (jumpToBuyPageDetails) {
+              setState(() {
+                StateContainer.of(context).updateTabPosition(tabPosition: 1);
+              });
+            }
+
+            positionStream =
+                Geolocator.getPositionStream().listen((Position position) {
+                  /* compare current and old position */
+                  if (position?.latitude != null &&
+                      tmpLocation?.latitude != null &&
+                      (position.latitude * 100).round() ==
+                          (tmpLocation!.latitude! * 100).round() &&
+                      (position.longitude * 100).round() ==
+                          (tmpLocation!.longitude * 100).round()) {
+                    widget.samePositionCount++;
+                  } else {
+                    widget.samePositionCount = 0;
+                    tmpLocation = StateContainer.of(context).location;
+                    if (position != null && mounted) {
+                      widget.hasGps = true;
+                      setState(() {
+                        StateContainer.of(context)
+                            .updateLocation(location: position);
+                      });
+                    }
+                  }
+                  if (widget.samePositionCount >= 3 || widget.hasGps!)
+                    positionStream?.cancel();
+                });
+          }
+        }
+      }
     });
   }
